+++ providers/capvcd/upgrade/capvcd_test.go
@@ -1,6 +1,8 @@
 package upgrade
 
 import (
+	"time"
+
 	. "github.com/onsi/ginkgo/v2"
 
 	"github.com/giantswarm/cluster-test-suites/internal/common"
@@ -8,7 +10,13 @@ import (
 )
 
 var _ = Describe("Basic upgrade test", Ordered, func() {
-	upgrade.Run()
+	// it is better to get defaults at first and then customize
+	// further changes in defaults will be effective here.
+	cfg := upgrade.NewTestConfigWithDefaults()
+	cfg.ControlPlaneNodesTimeout = 30 * time.Minute
+	cfg.WorkerNodesTimeout = 30 * time.Minute
+
+	upgrade.Run(cfg)
 
 	// Finally run the common tests after upgrade is completed
 	common.Run(&common.TestConfig{