+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add test config for upgrades to be able to customize timeouts per provider.
+
 ## [1.20.2] - 2023-12-13
 
 - Disable Bastion tests for `capa` provider.